Updated enemies
-Fixed an issue where the turret gun wasnt pointing at enemy
-Fixed the collision issue on enemies
- Made it so enemies can damage the player
- Applied textures
- Made the drones patrols look less like teleporting and more like actual patrolling.

A few things to note:
The turrets pawn sensing has been overhauled due to UE4s inability to rotate the pawn sensing component. This has resulted in the turret being much smoother overall but being a tad bit smarter than it should be.

The turrets texture is currently simply steel. This is because in order to have the rotating gun, the turret had to be split into different pieces which made applying the singular texture/UV map basically impossible.